Removed version number from built output.
The idea is that, due to backward compatibility, one should be
able to just drop the new library in place of the old one without
having to change the link argument.  True, this already kind of
happened with symbolic links and auto_link, but let's make it
even more explicit.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/branches/R01000x-branch/QuantLib@16794 8618b1d8-e22c-0410-b026-96a5dba3e089