Bug fix: QuantLib now compiles with Sun Studio on Solaris.
Unlike gcc and VC++, Sun Studio does not make math functions available
in the global namespace when including the <cmath> header.  Now we added
a few using directives for them, thus reproducing the other compilers'
bad behavior.  Not extremely clean, but it sure beats hunting the sources
for all instances of unqualified math functions before each release.

Other minor fixes include reordering a couple of expressions so that
they're more palatable to Sun Studio (for which v.end(), where v is
a std::vector, is not an lvalue) and adding the std:: qualification
to a function call.

Thanks to Norbert Irmer for the report and for testing the changes.

git-svn-id: https://quantlib.svn.sourceforge.net/svnroot/quantlib/branches/R01000x-branch/QuantLib@17250 8618b1d8-e22c-0410-b026-96a5dba3e089